ConcurrentParameterSet blocks until thread started
It is still difficult to guarantee that there will be no concurrency
problems while starting the thread (see the WARNING) comment in the
asyncDispatcherCallback function for more details), but at least now
starting the thread can be done in a blocking fashion much more safely.
This also means that SLEEP_AFTER_CREATION_MS is no longer needed.